+++ examples/google/main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -23,7 +22,7 @@ const (
 )
 
 // sessionStore encodes and decodes session data stored in signed cookies
-var sessionStore = sessions.NewCookieStore([]byte(sessionSecret), nil)
+var sessionStore = sessions.NewCookieStore(sessions.DebugCookieConfig, []byte(sessionSecret), nil)
 
 // Config configures the main ServeMux.
 type Config struct {
@@ -62,9 +61,12 @@ func issueSession() http.Handler {
 		}
 		// 2. Implement a success handler to issue some form of session
 		session := sessionStore.New(sessionName)
-		session.Values[sessionUserKey] = googleUser.Id
-		session.Values[sessionUsername] = googleUser.Name
-		session.Save(w)
+		session.Set(sessionUserKey, googleUser.Id)
+		session.Set(sessionUsername, googleUser.Name)
+		if err := session.Save(w); err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
 		http.Redirect(w, req, "/profile", http.StatusFound)
 	}
 	return http.HandlerFunc(fn)
@@ -75,12 +77,12 @@ func profileHandler(w http.ResponseWriter, req *http.Request) {
 	session, err := sessionStore.Get(req, sessionName)
 	if err != nil {
 		// welcome with login button
-		page, _ := ioutil.ReadFile("home.html")
-		fmt.Fprintf(w, string(page))
+		page, _ := os.ReadFile("home.html")
+		fmt.Fprint(w, string(page))
 		return
 	}
 	// authenticated profile
-	fmt.Fprintf(w, `<p>You are logged in %s!</p><form action="/logout" method="post"><input type="submit" value="Logout"></form>`, session.Values[sessionUsername])
+	fmt.Fprintf(w, `<p>You are logged in %s!</p><form action="/logout" method="post"><input type="submit" value="Logout"></form>`, session.Get(sessionUsername))
 }
 
 // logoutHandler destroys the session on POSTs and redirects to home.